refactor(frontend): extract useSidebarContextMenu hook from EditorLayout
